Fixed bug in AnaTreeMC; Renamed some files.
Fixed a bug which caused compilation to fail in AnaTreeMC. Shame on me
for pushing a commit which did not compile.

Renamed applications to fit naming scheme. All executables are now prefixed
with "xsllh". CMakeLists.txt has been updated with the name changes and
the instructions to compile ccqefit, fitbase, and propError have been
removed since they are not needed.

+++ src/xsllhDetVariations.cxx
@@ -255,8 +255,8 @@ int main(int argc, char** argv)
     //std::vector<bool> use_sample = {true, true, true, true, false, true, true, false, false, false};
     //const unsigned int num_elements = nbins * std::count(use_sample.begin(), use_sample.end(), true);
     const unsigned int num_elements = nbins * use_samples.size();
-    TMatrixTSym<float> cov_mat(num_elements);
-    TMatrixTSym<float> cor_mat(num_elements);
+    TMatrixTSym<double> cov_mat(num_elements);
+    TMatrixTSym<double> cor_mat(num_elements);
     cov_mat.Zero();
     cor_mat.Zero();
